refactor(admin): unifica Membri+Richieste in un tab, Ricorsi in tab dedicato
Il Pannello Amministratore aveva tab separati 'Gestione Membri' e
'Richieste', mentre la sezione 'Gestione Ricorsi' era un blocco
collassabile in fondo ai membri. Ora:

- 'Gestione Membri' unifica membri attivi + richieste di adesione +
  inviti (rende AdminMembersTab + AdminRequestsTab insieme).
- 'Ricorsi' e' un tab dedicato (nuovo AdminAppealsTab, estratto da
  AdminMembersTab) che gestisce i ricorsi di tutte le fasi d'asta.
- I link 'gestisci ricorso' da Svincolati/Rubata (tab: 'appeals')
  atterrano ora sul tab dedicato invece che nella sezione collassata.
- Badge: richieste pendenti sul tab Membri, ricorsi pendenti sul tab
  Ricorsi. mapTab retrocompatibile (requests/invites -> members).

Solo wiring/markup UI, nessuna logica backend toccata. Test aggiornati.
